requirements.txt
tqdm==4.64.0
opencv-python==4.6.0.66
h5py==3.6.0
numpy==1.24.4
matplotlib
pyyaml
rospkg
torch==1.11.0
torchvision==0.12.0
scipy==1.10.1
yq
vidgear
albumentations
